# BackendChatAplication/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mi API",
        "description": "Documentación de la API de usuarios",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/messages": {
            "get": {
                "tags": [
                    "Mensajes"
                ],
                "summary": "Mostrar lista de mensajes",
                "operationId": "9d5effa2cb648df85be0ea9741ffb457",
                "responses": {
                    "200": {
                        "description": "Lista de mensajes"
                    }
                }
            },
            "post": {
                "tags": [
                    "Mensajes"
                ],
                "summary": "Crear un nuevo mensaje",
                "operationId": "db79bb7bcadfdb708e9f6c38ac928f4a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "sender_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "receiver_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Hola, ¿cómo estás?"
                                    },
                                    "images[]": {
                                        "description": "Archivos de imagen",
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Mensaje creado"
                    }
                }
            }
        },
        "/api/messages/{id}": {
            "get": {
                "tags": [
                    "Mensajes"
                ],
                "summary": "Mostrar un mensaje específico",
                "operationId": "8938bb52056aef543d340164d37f9562",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mensaje específico"
                    }
                }
            },
            "put": {
                "tags": [
                    "Mensajes"
                ],
                "summary": "Actualizar un mensaje",
                "operationId": "a1b0a345721451dac454c4049d504775",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "content": {
                                        "type": "string",
                                        "example": "Hola, ¿cómo estás?"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mensaje actualizado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Mensajes"
                ],
                "summary": "Eliminar un mensaje",
                "operationId": "2077fda8b81cc2fb355890827a1ea31c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mensaje eliminado"
                    }
                }
            }
        },
        "/api/messages/between/{userId1}/{userId2}": {
            "get": {
                "tags": [
                    "Mensajes"
                ],
                "summary": "Obtener mensajes entre dos usuarios",
                "operationId": "0b00ec5a3d5ce3badd735e0084824881",
                "parameters": [
                    {
                        "name": "userId1",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "userId2",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mensajes entre usuarios"
                    }
                }
            }
        },
        "/api/messages/senders/{userId}": {
            "get": {
                "tags": [
                    "Mensajes"
                ],
                "summary": "Obtener remitentes de un usuario",
                "operationId": "3a81c4194e4ca4200906b8e55e4f4895",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de remitentes"
                    }
                }
            }
        },
...
    },
...
}